fix: align employee location migration key type

+++ database/migrations/2026_03_13_090000_create_employee_locations_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
 
         Schema::create('employee_locations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('employee_id');
+            $table->unsignedInteger('employee_id');
             $table->decimal('latitude', 10, 7);
             $table->decimal('longitude', 10, 7);
             $table->timestamp('timestamp')->index();
